Returns user-specified Expert Query National Extracts of ATTAINS data for use cases where nationwide data is desired.

This function is also useful in situations where the the desired query would yield more than 1 million rows as the national extracts can be sorted and filtered after import. The National Extracts are large files and can take several minutes to download and import.

Usage

EQ_NationalExtract(extract = NULL, max_retries = 3)

Arguments

extract

Character argument. Specifies which Expert Query National Extract should be imported. Options are "actions" (Actions), "assessments" (Assessments), "au" (Assessment Units), "au_mls" (Assessment Units with Monitoring Locations), "catch_corr" (Catchment Correspondence), "sources" (Sources), and "tmdl" (TMDLs). There is no national extract option available for Actions Documents. The default is NULL which means no extract will be returned.

max_retries

Integer. The number of retry attempts.

Value

A data frame containing the user-specified national extract. The columns returned will vary based on the extract selected and are as follows:

"actions" (Actions): "objectId", "region", "state", "organizationType", "organizationId", "organizationName", "waterType", "parameterGroup", "parameter", "actionType", "actionId", "actionName", "actionAgency", "inIndianCountry", "includeInMeasure", "completionDate", "assessmentUnitId", "assessmentUnitName", "fiscalYearEstablished", "locationDescription". "waterSize", "waterSizeUnits", and "planSummaryLink".

"assessments" (Assessments): "objectId", "region", "state", "organizationType", "organizationId", "organizationName", "waterType", "reportingCycle", "cycleLastAssessed", "assessmentUnitId", "assessmentUnitName", "assessmentUnitStatus", "overallStatus", "epaIrCategory", "stateIrCategory", "useGroup", "useName", "useClassName", "useSupport", "useIrCategory", "useStateIrCategory", "monitoringStartDate", "monitoringEndDate", "assessmentDate", "assessmentTypes", assessmentMethods", "assessmentBasis", "parameterGroup", "parameterName", "parameterStatus", "parameterAttainment", parameterIrCategory" "parameterStateIrCategory", delisted", "delistedReason", "pollutantIndicator", "cycleFirstListed", "alternateListingIdentifier", "vision303dPriority", "cwa303dPriorityRanking", "cycleScheduledForTmdl", "cycleExpectedToAttain", "consentDecreeCycle", "cycleId", "seasonStartDate", "seasonEndDate", "associatedActionId", "associatedActionName", "associatedActionType", "associatedActionStatus", "associatedActionAgency", "locationDescription", "sizeSource", "sourceScale", "waterSize", and "waterSizeUnits".

"aus" (Assessment Units): "objectId", "region", "state", "organizationType", "organizationId", "organizationName", "useClassName", "assessmentUnitId", "assessmentUnitName", "assessmentUnitStatus", "reportingCycle", "cycleId", "locationDescription", "sizeSource", "sourceScale", "waterSize", and "waterSizeUnits".

"au_mls" (Assessment Units with Monitoring Locations): "objectId", "region", "state", "organizationType", "organizationId", "organizationName", "waterType", "useClassName", "monitoringLocationId", "monitoringLocationOrgId", "assessmentUnitId", "assessmentUnitName", "assessmentUnitStatus", "reportingCycle", "cycleId", locationDescription", "monitoringLocationDataLink", "sizeSource", "sourceScale", "waterSize", and "waterSizeUnits"

"catch_corr" (Catchment Correspondence): "objectId", "region", "state", organizationType", "organizationId", "organizationName", "assessmentUnitId", "assessmentUnitName", "catchmentNhdPlusId", "reportingCycle", and "cycleId".

"sources" (Sources): "objectId", "region", "state", "organizationType", "organizationId", "organizationName", "waterType", "assessmentUnitId", "assessmentUnitName", "reportingCycle", "overallStatus", "epaIrCategory", "stateIrCategory", "parameterGroup", "causeName", "sourceName", "confirmed", "cycleId", "locationDescription", "waterSize", and "waterSizeUnits".

"tmdl" (TMDLs): "objectId", "region", "state", "organizationType", "organizationId", "organizationName", "waterType", "pollutantGroup", "pollutant", "addressedParameterGroup", "addressedParameter", "sourceType", "npdesIdentifier", "otherIdentifier", "actionId", "actionName", "actionAgency", "inIndianCountry", "explicitMarginOfSafety", implicitMarginOfSafety", "includeInMeasure", "completionDate", "tmdlDate", "fiscalYearEstablished", "assessmentUnitId", "assessmentUnitName", "loadAllocation", "loadAllocationUnits", "locationDescription", "tmdlEndpoint", "waterSize", "waterSizeUnits", "wasteLoadAllocation", and "planSummaryLink".

Details

National extracts can and more information about Expert Query can be found here: https://owapps.epa.gov/expertquery/national-downloads

If using VPN, suggest signing out of VPN before running this function as it can cause 403 error and prevent download of national extracts.

** The National Extracts are large files. You must have enough memory available in order for these functions to import them into R successfully.

HRM NOTE 2/11/25 - can download but not import catchment correspondence due to file size.

Examples

if (FALSE) { # \dontrun{
assessments <- EQ_NationalExtract(extract = "assessments")

aus_monloc <- EQ_NationalExtract(extract = "au_mls")
} # }